e2etest: move dns from validator-tests
This commit moves interface of dns actor from validator-test into e2etest-dns. It
also updates crate references.
